Title: 이차원 배열의 오름차순 정리, Time: 39ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240829/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.cpp
@@ -9,7 +9,7 @@ int main() {
     while(left<=right){
         long long mid=(left+right)/2;
 
-        int cnt=0;
+        long long cnt=0;
         for(int i=1; i<=n; ++i){
             cnt+=min(n, mid/i);
         }
